Support changing other properties besides brightness
Add support for changing other properties besides brightness.  We can now change the step delay and the precision of the hue change. This allows us to speed up the rainbow effect in various ways (or, potentially, to slow it down).

New constructors were added for default brightness, brightness+delay, and brightness+delay+precision.  This change allows for including multiple different rainbow effects with different parameters in a single sketch without having to adjust each parameter individually for each effect object.

Finally, reset the default brightness to 151 to make the plugin default values match the values used in the default Model01 firmware sketch.  The older value produced a lackluster-effect that needed to be updated in the sketch to look decent; this change means that default setting can be removed from the default firmware sketch.